requirements.txt
flask==1.1.2
imageio==2.5.0
tensorflow==2.1.0
keras==2.3.0
pylint==2.6.0
google-cloud-storage==1.30.0
